Move value unwrapping back to the producer side
This closes #15; see https://github.com/tc39/proposal-async-iteration/issues/15#issuecomment-247732544. It ensures that async generators always (*) produce objects which conform to the async iterator contract, namely in not producing Promise<IterationResult<Promise<T>>> from their next() methods. It then removes the double-unwrapping from for-await-of, as for-await-of should be able to assume that it is dealing with a contract-following producer. As a bonus, this makes for-await-of consistent with yield* (in that now neither do double-unwrapping).

This involves much fewer changes than https://github.com/tc39/proposal-async-iteration/commit/f03e40e1484c365b8769b9faed0277f19255a602, which did the opposite by moving the unwrapping from the consumer to the producer side. This is largely because that commit also was removing double-unwrapping in yield* at the same time that it added it to for-await-of.

(*): "always" caveat: if you yield* an async iterator which does not follow the contract, you will not follow the contract.